+++ build_runner/lib/src/commands/watch_command.dart
@@ -50,6 +50,10 @@ class WatchCommand implements BuildRunnerCommand {
       b.onLog = testingOverrides.onLog;
     });
 
+    if (buildOptions.workspace) {
+      buildLog.warnAboutWorkspaceFlag();
+    }
+
     final buildPlan = await BuildPlan.load(
       builderFactories: builderFactories,
       buildOptions: buildOptions,